Refactor transform trace API helpers
